requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pandas==1.2.3
scikit-learn
Werkzeug
numpy
detecto
opencv-contrib-python-headless
vietocr==0.3.5
matplotlib
Pillow
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu 
torchvision==0.9.1+cpu
pyyaml